# static/swagger.yaml
openapi: 3.0.3
info:
  title: Kimiko Golang Swagger
  description: |-
    This API provides functionalities for user management, organisation management, authentication, messaging, payments, and various other services.
    It includes features like social authentication, email notifications, payment integrations with Stripe and Flutterwave, and comprehensive user and organisation management.
  termsOfService: http://example.com/terms/
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  # Added by API Auto Mocking Plugin
  - url: https://staging.api-golang.boilerplate.hng.tech/api/v1
  - url: https://deployment.api-golang.boilerplate.hng.tech/api/v1
...
paths:
...
  /jobs:
    get:
      tags:
        - Job Posts
      summary: Fetch all job posts
      description: Fetches all job posts with optional pagination.
      operationId: fetchAllJobPosts
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
          description: The page number for pagination
          example: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
          description: The number of items per page
          example: 6
      responses:
        '200':
          description: A list of paginated job posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobPost'
        '404':
          description: Jobs not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Job Posts
      summary: Create a new job post
      description: Creates a new job post with the provided details.
      operationId: createJobPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobPost'
      responses:
        '201':
          description: Job post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPostResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to create job post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs/{job_id}:
    get:
      tags:
        - Job Posts
      summary: Fetch a job post by ID
      description: Fetches the details of a job post by its ID.
      operationId: fetchJobPostByID
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the job post to fetch
      responses:
        '200':
          description: Job post fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPostResponse'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to fetch job post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Job Posts
      summary: Update a job post by ID
      description: Updates the details of an existing job post.
      operationId: updateJobPostByID
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the job post to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobPost'
      responses:
        '200':
          description: Job post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPostResponse'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to update job post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Job Posts
      summary: Delete a job post by ID
      description: Deletes a job post by its ID.
      operationId: deleteJobPostByID
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the job post to delete
      responses:
        '204':
          description: No content
                  
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to delete job post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateJobPost:
      type: object
      properties:
        title:
          type: string
        salary_range:
          type: string
        job_type:
          type: string
        location:
          type: string
        deadline:
          type: string
        job_mode:
          type: string
        experience_level:
          type: string
        benefits:
          type: string
        company_name:
          type: string
        description:
          type: string
        key_responsibilities:
          type: string
        qualifications:
          type: string
    UpdateJobPost:
      type: object
      properties:
        title:
          type: string
          example: Updated Software Engineer Intern
        salary_range:
          type: string
          example: 6000-8000 USD
        job_type:
          type: string
          example: full-time
        location:
          type: string
          example: San Francisco, CA
        deadline:
          type: string
          format: date-time
          example: 2024-12-31T23:59:59Z
        job_mode:
          type: string
          example: hybrid
        experience_level:
          type: string
          example: Entry level (0-3 years)
        benefits:
          type: string
          example: Flexible hours, Health insurance, Stock options
        company_name:
          type: string
          example: Tech Innovators Inc.
        description:
          type: string
          example: We are looking for a passionate Software Engineer Intern to join our team. You'll work on exciting projects and gain valuable experience.
        key_responsibilities:
          type: string
          example: Develop and maintain web applications, Collaborate with the team on various projects, Participate in code reviews
        qualifications:
          type: string
          example: Ability to work in a team, Bachelor degree in Computer Science
    JobPost:
      type: object
      properties:
        id:
          type: string
          example: 01910544-d1e1-7ada-bdac-c761e527ec91
        title:
          type: string
          example: Junior Backend intern
        salary_range:
          type: string
          example: 500,000 NGN
        job_type:
          type: string
          example: internship
        location:
          type: string
          example: San Francisco, CA
        deadline:
          type: string
          format: date-time
          example: 2024-12-31T23:59:59Z
        job_mode:
          type: string
          example: remote
        experience_level:
          type: string
          example:  0-2 years
        how_to_apply:
          type: string
          example: Submit your resume and cover letter to hr@company.com
        benefits:
          type: string
          example: Flexible hours, Remote work, Health insurance
        company_name:
          type: string
          example: Tech Innovators
        description:
          type: string
          example: We are looking for a passionate Software Engineer Intern to join our team. You will be working on exciting projects and gain hands-on experience.
        key_responsibilities:
          type: string
          example: Develop and maintain web applications, Collaborate with the team on various projects, Participate in code reviews
        qualifications:
          type: string
          example: Expert with GIT
        created_at:
          type: string
          format: date-time
          example: 2024-07-30T21:11:21.9538358+01:00
        updated_at:
          type: string
          format: date-time
          example: 2024-07-30T21:11:21.9538358+01:00
    JobPostResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        status_code:
          type: integer
          example: 201
        message:
          type: string
          example: Job created successfully
        data:
          $ref: '#/components/schemas/JobPost'
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        status_code:
          type: integer
          example: 400
        message:
          type: string
          example: An error occurred
...
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT